Excel_Project/requirements.txt
asgiref==3.8.1
Django==4.2.15
djangorestframework==3.15.2
et-xmlfile==1.1.0
openpyxl==3.1.5
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
